# src/Data_Module/openapi.yaml
openapi: 3.1.0
info:
  title: Data Module API
  description: Intelligent Logistics - Data Module (Source of Truth)
  version: 1.0.0
servers:
  - url: http://10.255.32.82:8080/api/v1
    description: Production server
paths:
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List Alerts
      description: Lists alerts with optional filters.
      operationId: list_alerts
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: alert_type
          in: query
          schema:
            type: string
            nullable: true
        - name: visit_id
          in: query
          schema:
            type: integer
            nullable: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
    post:
      tags: [Alerts]
      summary: Create Manual Alert
      description: Creates an alert manually.
      operationId: create_manual_alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /alerts/active:
    get:
      tags: [Alerts]
      summary: List Active Alerts
      description: Lists active alerts (last 24h).
      operationId: list_active_alerts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
...
  /alerts/visit/{visit_id}:
    get:
      tags: [Alerts]
      summary: Get Visit Alerts
      description: Gets all alerts for a specific visit.
      operationId: get_visit_alerts
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
  /alerts/{alert_id}:
    get:
      tags: [Alerts]
      summary: Get Single Alert
      description: Gets details of a specific alert.
      operationId: get_single_alert
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
...
  /alerts/hazmat:
    post:
      tags: [Alerts]
      summary: Create Hazmat ADR Alert
      description: Creates hazmat/ADR specific alert.
      operationId: create_hazmat_alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHazmatAlertRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'

  # ==================== WORKERS ====================
...
components:
  schemas:
...
    Alert:
      type: object
      required: [id, timestamp]
      properties:
        id:
          type: integer
        visit_id:
          type: integer
          nullable: true
        type:
          $ref: '#/components/schemas/TypeAlertEnum'
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    TypeAlertEnum:
      type: string
      enum: [generic, safety, problem, operational]
    CreateAlertRequest:
      type: object
      required: [type, description]
      properties:
        visit_id:
          type: integer
          nullable: true
        type:
          type: string
        description:
          type: string
        image_url:
          type: string
          nullable: true
    CreateHazmatAlertRequest:
      type: object
      required: [appointment_id]
      properties:
        appointment_id:
          type: integer
        un_code:
          type: string
          nullable: true
        kemler_code:
          type: string
          nullable: true
        detected_hazmat:
          type: string
          nullable: true
...